Update typography.css
Reintroducing the comments around :root{}, .prose{} and .prose :where(code):not(:where([class~="not-prose"] *)){} classes because the lack of comments causes improper rendering of markdown code blocks as HTML on mobile Safari (and also on desktop Firefox/Chrome if the window width is extremely small).  Probably the only one of the three classes that actually needs to be commented to fix the improper rendering of code blocks is the third one (.prose :where(code):not(:where([class~="not-prose"] *)){}) but in my opinion the site renders correctly without any of the three overrides.

Commenting out .prose {max-width: 65ch;} (line #3 in the non-minified file) because this forces a left-of-center alignment of body content and causes the body content to seem arbitrarily less wide than the banner image.