Update MarkdownPagesBase.cs
Fixing the header anchor links so that they actually work (instead of just invoking an empty onclick event).  This allows both for a user to click on a header text item to scroll the page to that location as well as for the user to right-click on the anchor link and copy it to the clipboard.